# Use stream instead of parallelStream