--- a/buildSrc/src/main/groovy/com/uber/okbuck/composer/base/BuckRuleComposer.java
+++ b/buildSrc/src/main/groovy/com/uber/okbuck/composer/base/BuckRuleComposer.java
@@ -14,13 +14,13 @@ public abstract class BuckRuleComposer {
     private static final String SEPARATOR = ":";
 
     public static Set<String> external(final Set<String> deps) {
-        return deps.parallelStream()
+        return deps.stream()
                 .map(BuckRuleComposer::fileRule)
                 .collect(Collectors.toSet());
     }
 
     public static Set<String> externalApt(final Set<String> deps) {
-        return external(deps).parallelStream()
+        return external(deps).stream()
                 .filter(dep -> dep.endsWith(".jar"))
                 .collect(Collectors.toSet());
     }
@@ -40,13 +40,13 @@ public abstract class BuckRuleComposer {
     }
 
     public static Set<String> targets(final Set<Target> deps) {
-        return deps.parallelStream()
+        return deps.stream()
                 .map(BuckRuleComposer::targets)
                 .collect(Collectors.toSet());
     }
 
     public static Set<String> targetsApt(final Set<Target> deps) {
-        return deps.parallelStream()
+        return deps.stream()
                 .filter(target -> target.getClass().equals(JavaLibTarget.class))
                 .map(BuckRuleComposer::targets)
                 .collect(Collectors.toSet());
@@ -61,7 +61,7 @@ public abstract class BuckRuleComposer {
     }
 
     public static String toLocation(final List<String> targets) {
-        return targets.parallelStream()
+        return targets.stream()
                 .map(BuckRuleComposer::toLocation)
                 .collect(Collectors.joining(SEPARATOR));
     }