--- a/buildSrc/src/main/java/com/uber/okbuck/core/util/LintUtil.java
+++ b/buildSrc/src/main/java/com/uber/okbuck/core/util/LintUtil.java
@@ -36,7 +36,7 @@ public final class LintUtil {
                 .getByName("classpath")
                 .getResolvedConfiguration()
                 .getResolvedArtifacts()
-                .parallelStream()
+                .stream()
                 .filter(resolvedArtifact -> {
                     ModuleVersionIdentifier identifier = resolvedArtifact.getModuleVersion().getId();
                     return (LINT_GROUP.equals(identifier.getGroup()) && LINT_MODULE.equals(identifier.getName()));