The overhead of creating threads for very small tasks might not be
faster than performing the same tasks serially.

I ran a benchmark using this commit against our repo. I killed the
gradle daemon and then ran the okbuck task 3 times.

Before: using parallel()
20.353 secs
6.644 secs
5.89 secs

After: using stream()
18.539 secs
6.316 secs
5.613 secs

The results were quite similar with and without parallel stream. I'm
wondering if there is any noticeable difference with even larger projects.