@kageiit 2017-05-04
Ran it on our codebase

Before:
```
Total time: 53.786 secs
Total time: 34.565 secs
Total time: 32.101 secs
```

After
```
Total time: 51.386 secs
Total time: 34.451 secs
Total time: 30.859 secs
```

Looks good