@runningcode 2017-05-04
Removed snapshot version!